requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.34.0
boto3==1.35.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
bcrypt==4.0.1
pyjwt==2.8.0